startup message

+++ GUI/MainWindow.xaml.cs
@@ -13,6 +13,7 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
+using System.Windows.Documents;
 using System.Windows.Input;
 using TaskLayer;
 
@@ -70,14 +71,14 @@ namespace MetaMorpheusGUI
 
             // LOAD GUI SETTINGS
 
-            try
+            if (File.Exists(@"GUIsettings.toml"))
             {
                 GuiGlobalParams = Toml.ReadFile<GuiGlobalParams>(Path.Combine(GlobalVariables.DataDir, @"GUIsettings.toml"));
-                notificationsTextBox.Document.Blocks.Clear();
-            } catch (FileNotFoundException)
+            }
+            else
             {
-                // first time running
                 Toml.WriteFile(GuiGlobalParams, Path.Combine(GlobalVariables.DataDir, @"GUIsettings.toml"), MetaMorpheusTask.tomlConfig);
+                notificationsTextBox.Document = YoutubeWikiNotification();
             }
 
             if (GlobalVariables.MetaMorpheusVersion.Contains("Not a release version"))
@@ -91,6 +92,42 @@ namespace MetaMorpheusGUI
             {
                 GuiWarnHandler(null, new StringEventArgs("Could not get newest version from web: " + e.Message, null));
             }
+        }
+
+        private FlowDocument YoutubeWikiNotification()
+        {
+            FlowDocument doc = new FlowDocument();
+            Paragraph p = new Paragraph();
+            Run run1 = new Run("Visit our ");
+            Run run2 = new Run("Wiki");
+            Run run3 = new Run(" or ");
+            Run run4 = new Run("Youtube");
+            Run run5 = new Run(" for more information.");
+
+            Hyperlink wikiLink = new Hyperlink(run2);
+            wikiLink.NavigateUri = new Uri(@"https://github.com/smith-chem-wisc/MetaMorpheus/wiki");
+
+            Hyperlink youtubeLink = new Hyperlink(run4);
+            youtubeLink.NavigateUri = new Uri(@"https://www.youtube.com/playlist?list=PLVk5tTSZ1aWlhNPh7jxPQ8pc0ElyzSUQb");
+
+            var links = new List<Hyperlink> {wikiLink, youtubeLink};
+
+            p.Inlines.Add(run1);
+            p.Inlines.Add(wikiLink);
+            p.Inlines.Add(run3);
+            p.Inlines.Add(youtubeLink);
+            p.Inlines.Add(run5);
+
+            foreach (Hyperlink link in links)
+            {
+                link.RequestNavigate += (sender, e) =>
+                {
+                    System.Diagnostics.Process.Start(e.Uri.ToString());
+                };
+            }
+
+            doc.Blocks.Add(p);
+            return doc;
         }
 
         public static string NewestKnownVersion { get; private set; }
@@ -479,7 +516,7 @@ namespace MetaMorpheusGUI
                             try
                             {
                                 GlobalVariables.AddMods(UsefulProteomicsDatabases.ProteinDbLoader.GetPtmListFromProteinXml(draggedFilePath).OfType<ModificationWithLocation>());
-                                
+
                                 PrintErrorsReadingMods();
                             }
                             catch (Exception ee)
@@ -787,7 +824,7 @@ namespace MetaMorpheusGUI
                 UpdateTaskGuiStuff();
             }
         }
-        
+
         // deletes the selected task
         private void DeleteSelectedTask(object sender, RoutedEventArgs e)
         {
@@ -1352,17 +1389,6 @@ namespace MetaMorpheusGUI
                 AddAFile(contaminantFile);
             }
             dataGridProteinDatabases.Items.Refresh();
-        }
-
-        // add wiki and youtube links to notifications box
-        private void WikiHyperlink(object sender, System.Windows.Navigation.RequestNavigateEventArgs e)
-        {
-            System.Diagnostics.Process.Start(@"https://github.com/smith-chem-wisc/MetaMorpheus/wiki");
-        }
-
-        private void YoutubeHyperlink(object sender, System.Windows.Navigation.RequestNavigateEventArgs e)
-        {
-            System.Diagnostics.Process.Start(@"https://www.youtube.com/playlist?list=PLVk5tTSZ1aWlhNPh7jxPQ8pc0ElyzSUQb");
         }
     }
 }
